Moved suggestions to the comp
